Enhanced DateTime Handling for supporting non ISO 8601 DateTimes
